Fix: Tox workflow takes only one python version
The previous version of this workflow was based on an incorrect
understanding of how GHA would process YAML maps for a matrix. This
was being used to specify different workflows for different python
versions.

Since this does not work as intended, we will now simply use a single
python version, and a list of envs. If the caller would like to run
workflows on multiple versions of python, they can still call the
workflow again on a different version.

This change also includes improvements to the documentation,
particularly focused on improving the user's understanding of how to
pass in the env list.

Issue: RELENG-5132
Signed-off-by: Eric Ball <eball@linuxfoundation.org>