Fix: Remove "cache: pip" from setup-python call
This is a convenient way to get requirements installed, but requires
either requirements.txt or pyproject.toml to be present. Since we
don't want the action to fail if these are not present, we can't use
it here. Instead, the build-package step will install pyproject.toml
requirements, and we will now check for a requirements.txt file while
running the shell script that calls tox.

This also adds pipx to the tox call, to ensure that tox and its
requirements are present to run tox properly.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>