Fix: Only obtain global-settings through variable
Currently, the global-settings are obtained through a Jenkins fetch,
and then via a variable which is concatenated onto what was fetched.
This is a problem, as it will create an invalid settings file if both
are defined.

To simplify this process, we can simply use the variable. The
touch/cat is replaced with an overwrite, as we want the file to be
written as it's stored, and not placed on top of anything that is
already in the file.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>