Fix: Use single quotes for global-settings echo
Because the variable is being filled by the action rather than bash,
using double quotes is messing with the double quotes that exist
within the text of the variable. By using single quotes, bash will
handle all text literally, and the variable will still dereference,
since that is happening before the script is handed to bash.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>