Feat: Add maven-merge workflow
This workflow is intended to build, test, and sign a java package,
before uploading the package as an artifact. It can then be
downloaded by another workflow and pushed to specified repos.

Because this is very project-specific, rather than creating a single
workflow that will likely only work for one project, this includes a
reference workflow that can be customized and implemented by each
individual project. This has been added to the docs directory since it
is documenting an example, rather than being a workflow that is
intended to be used exactly as-is.

Issue-ID: RELENG-5121
Signed-off-by: Eric Ball <eball@linuxfoundation.org>