Fix: gerrit-required-info-yaml-verify fetch-depth
As the point of the workflow is to verify changes to INFO.yaml,
and the logic skips the check if that file is not edited,
and "git diff --name-only HEAD~1" is used to detect edits,
the default fetch-depth of lfit/checkout-gerrit-change-action
(v0.9) of 1 is not enough.

Fix by applying check-depth 2,
also add a comment explaining why.

Signed-off-by: Vratko Polak <vrpolak@cisco.com>