Fix: Install setuptools for tox jobs with py3.12
The lftools license check fails the tox run with python 3.12.

The default version of the python used in the tox workflow
using python 3.12. Python 3.12 does not come with a stdlib
distutils module (changelog), because distutils was deprecated
in 3.10 and removed in 3.12. See PEP 632 – Deprecate distutils
module.

Therefore, check if python 3.12 and install distutils before the
run.

Ref:
https://docs.python.org/3/whatsnew/3.12.html
https://peps.python.org/pep-0632/
Issue: RELENG-5403
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>