CI: Enforce semantic commit messages
The semantic commit message validator we were (and still are using) only
validates that the PR matches semantic commit message standard and if
there is a single commit in the PR, that it also matches the PR title.

Since we require that _all_ of our commits meet the standard we needed
to find another way to do this.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>