Fix: Use the Maven pom file provided by input
While there is an input parameter for setting a custom pom.xml file
there is no code to actually utilize it. This change fixes that
issue and enables the use of alternative pom.xml file path. Also
fix the description of parameter to state that its the path to the
file and not directory.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>