Test: Create minimal unit test for Maven jobs
This change creates a minimal pom.xml for unit testing of Maven jobs
and creates a test job for the Maven Verify workflow. The intent here
is to create a small unit test that can quickly validate Maven
workflows and catch potential issues that maybe missed by a linter.

This change also modifies the compose-maven-verify job to accept a new
parameter MVN_GLOBAL_SETTINGS which is optional and defaults the the
current behaviour while allowing it to be overrided to provide an
alternative global-settings which is necessary for this unit test to
work on PRs from forked repos.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>