Feat: Rework standalone linting action
Download of additional dependencies is now optional and there is no default
URL. Pre-commit is installed separately and it will automatically download
whatever the dependencies are for MyPy, given it installs linters in a venv.
The python version is now an optional input, so the hard coded value is now
at the top of the file. The /tmp directory is now used for the requirements
file when it is downloaded. I've updated the reusable workflow example but
removed in from the README.md documentation.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>