Fix: Remove remnants of compound URL handling
A bug was introduced at the time this action was imported/migrated over from
OS Climate. Previously, it was used to build compound URLs for checking PyPI
packages. That functionality was replaced, and this action should be simplfied.
Any compound handling of URLs should be performed elsewhere prior to calling
this action.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>