Chore: Update linting to allow merging of automation PRs
Update the .gitlint setup file to permit long first commit lines.

Also need to set this false in the verify.yaml workflow:
          validateSingleCommitMatchesPrTitle: false

This is blocking dependabot PRs, which sometimes contain the version
string for the updates applied in the PR title, which doesn't match
the first line of the commit message.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>